Simplified general user script request page and added create script page for admin
